+++ config/model_settings.php
@@ -2,4 +2,5 @@
 
 return [
     'settings_field_name' => env('MODEL_SETTINGS_FIELD_NAME', 'settings'),
+    'settings_table_name' => env('MODEL_SETTINGS_TABLE_NAME', 'model_settings'),
 ];