+++ src/Managers/FieldSettingsManager.php
@@ -4,16 +4,29 @@ namespace Glorand\Model\Settings\Managers;
 
 use Glorand\Model\Settings\Contracts\SettingsManagerContract;
 
+/**
+ * Class FieldSettingsManager
+ * @package Glorand\Model\Settings\Managers
+ * @property \Illuminate\Database\Eloquent\Model|\Glorand\Model\Settings\Traits\HasSettingsField $model
+ */
 class FieldSettingsManager extends AbstractSettingsManager
 {
+    /**
+     * @param array $settings
+     * @return \Glorand\Model\Settings\Contracts\SettingsManagerContract
+     */
     public function apply(array $settings = []): SettingsManagerContract
     {
-        $this->model->settings = $settings;
+        $this->model->{$this->model->getSettingsFieldName()} = json_encode($settings);
         $this->model->save();
 
         return $this;
     }
 
+    /**
+     * @param string|null $path
+     * @return \Glorand\Model\Settings\Contracts\SettingsManagerContract
+     */
     public function delete(string $path = null): SettingsManagerContract
     {
         if (!$path) {
@@ -28,6 +41,11 @@ class FieldSettingsManager extends AbstractSettingsManager
         return $this;
     }
 
+    /**
+     * @param string $path
+     * @param mixed $value
+     * @return \Glorand\Model\Settings\Contracts\SettingsManagerContract
+     */
     public function set(string $path, $value): SettingsManagerContract
     {
         $settings = $this->all();