+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to `glorand/laravel-model-settings` will be documented in this file
 
+## 3.0.0 - 2019-03-21
+### Added
+- Command to create the table for settings
+- Dynamic name for the settings table and field
+
+## 2.0.0 - 2019-01-26
+### Added
+- Console Commands
+
 ## 1.0.0 - 2019-01-08
 ### Added
 - Initial release